+++ src/lib/index.ts
@@ -3,3 +3,4 @@ export { KookeeApiError } from './http-client';
 export * from './types';
 export * from './fields';
 export * from './modules';
+export * from './seo';